Move WeatherAppViewModel to ui package
